Database refactoring
