+++ components/Tour/StageCard.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { DateFormat, formatDate, getDurationFormatted } from "@/utils/dateUtil";
 
-import { Stage } from "@/model/model";
+import { Stage } from "@/database/model/model";
 import { deleteStage } from "@/services/data/stageService";
 
 import {
