+++ components/Tour/TourList.tsx
@@ -1,4 +1,4 @@
-import { Tour } from "@/model/model";
+import { Tour } from "@/database/model/model";
 import { withObservables } from "@nozbe/watermelondb/react";
 import { Text, Card, Layout } from "@ui-kitten/components";
 import { Link } from "expo-router";
