+++ database/adapter.ts
@@ -1,0 +1,8 @@
+import SQLiteAdapters from "@nozbe/watermelondb/adapters/sqlite";
+import { schemas } from "./schemas";
+
+const adapter = new SQLiteAdapters({
+  schema: schemas,
+});
+
+export default adapter;
