+++ database/schemas/index.ts
@@ -1,0 +1,3 @@
+import { schema } from "./schema";
+
+export const schemas = schema;
